+++ lib/src/main/java/org/asamk/signal/manager/util/MessageCacheUtils.java
@@ -1,5 +1,7 @@
 package org.asamk.signal.manager.util;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.whispersystems.signalservice.api.messages.SignalServiceEnvelope;
 import org.whispersystems.signalservice.api.push.ServiceId;
 import org.whispersystems.signalservice.api.push.SignalServiceAddress;
@@ -15,11 +17,17 @@ import java.util.Optional;
 
 public class MessageCacheUtils {
 
+    private final static Logger logger = LoggerFactory.getLogger(MessageCacheUtils.class);
+
+    final static int CURRENT_VERSION = 8;
+
     public static SignalServiceEnvelope loadEnvelope(File file) throws IOException {
         try (var f = new FileInputStream(file)) {
             var in = new DataInputStream(f);
             var version = in.readInt();
-            if (version > 7) {
+            logger.trace("Reading cached envelope file with version {} (current: {})", version, CURRENT_VERSION);
+            if (version > CURRENT_VERSION) {
+                logger.warn("Unsupported envelope version {} (current: {})", version, CURRENT_VERSION);
                 // Unsupported envelope version
                 return null;
             }
@@ -71,6 +79,10 @@ public class MessageCacheUtils {
             if (version >= 7) {
                 isStory = in.readBoolean();
             }
+            String updatedPni = null;
+            if (version >= 8) {
+                updatedPni = in.readUTF();
+            }
             Optional<SignalServiceAddress> addressOptional = sourceServiceId == null
                     ? Optional.empty()
                     : Optional.of(new SignalServiceAddress(sourceServiceId, source));
@@ -84,6 +96,7 @@ public class MessageCacheUtils {
                     uuid,
                     destinationUuid == null ? UuidUtil.UNKNOWN_UUID.toString() : destinationUuid,
                     isUrgent,
+                    updatedPni == null ? "" : updatedPni,
                     isStory);
         }
     }
@@ -91,7 +104,7 @@ public class MessageCacheUtils {
     public static void storeEnvelope(SignalServiceEnvelope envelope, File file) throws IOException {
         try (var f = new FileOutputStream(file)) {
             try (var out = new DataOutputStream(f)) {
-                out.writeInt(7); // version
+                out.writeInt(CURRENT_VERSION); // version
                 out.writeInt(envelope.getType());
                 out.writeUTF(""); // legacy number
                 out.writeUTF(envelope.getSourceUuid().isPresent() ? envelope.getSourceUuid().get() : "");
@@ -111,6 +124,7 @@ public class MessageCacheUtils {
                 out.writeLong(envelope.getServerDeliveredTimestamp());
                 out.writeBoolean(envelope.isUrgent());
                 out.writeBoolean(envelope.isStory());
+                out.writeUTF(envelope.getUpdatedPni() == null ? "" : envelope.getUpdatedPni());
             }
         }
     }